Make snippets/dependencyManagement/customizingResolution-attributeSubstitutionRule...
...compatible with the configuration cache.

The task's `doLast` action was reaching out to `configurations` property
of the Gradle script object, causing a reference to this object to be
captured by the action. The configuration cache was attempting to
serialize the action and everything it captures but was failing to do
that because the script object is not serializable.

To work around that a separate `FileCollection` variable `classpath` is
introduced, that is initialized with the requested configuration. The
action only captures this variable, which can be safely serialized by
the configuration cached. We cannot simply store the configuration
itself in the variable, as the `Configuration` object is not serializable
beacause of all extra stuff it has compared to the simpler
`FileCollection`.

The changed code is not present in the user documentation and serves
only as a scaffolding, so no changes to the docs are necessary.